requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
faiss-cpu>=1.7.4
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
pytest>=7.4.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0
streamlit>=1.30.0
gdown>=4.7.0
